Implement Bee k8s client with interfaces
